requirements.txt
nmap3
pandas
requests
streamlit
openpyxl
python-docx
apscheduler
pyinstaller
